Added Book with authorids and bookimages

+++ OnlineLibraryManagementSystem/Controllers/AdminBookController.cs
@@ -1,11 +1,14 @@
 ﻿using Humanizer.Localisation;
+using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure.Internal;
 using OnlineLibraryManagementSystem.Data;
 using OnlineLibraryManagementSystem.Models.Admin.Book;
+using OnlineLibraryManagementSystem.Models.Authentication.SignUp;
 using OnlineLibraryManagementSystem.ViewModels;
 using System.Net;
+using System.Xml.Linq;
 using static Microsoft.EntityFrameworkCore.DbLoggerCategory;
 
 namespace OnlineLibraryManagementSystem.Controllers
@@ -15,10 +18,12 @@ namespace OnlineLibraryManagementSystem.Controllers
     public class AdminBookController : ControllerBase
     {
         private readonly ApplicationDbContext _context;
+        private readonly IWebHostEnvironment _webHostEnvironment;
 
-        public AdminBookController(ApplicationDbContext context)
+        public AdminBookController(ApplicationDbContext context, IWebHostEnvironment webHostEnvironment)
         {
             _context = context;
+            _webHostEnvironment = webHostEnvironment;
         }
 
         //-------------------------------BOOKS-------------------------------
@@ -33,7 +38,8 @@ namespace OnlineLibraryManagementSystem.Controllers
                 var books = await _context.Books
                 .Include(b => b.BookAuthors!)
                 .ThenInclude(ba => ba.Author)
-                .Select(x => new BookVM
+                .Include(b => b.BookImages)
+                .Select(x => new
                 {
                     BookName = x.BookName!,
                     Genre = x.Genre!,
@@ -46,7 +52,8 @@ namespace OnlineLibraryManagementSystem.Controllers
                     Description = x.Description!,
                     ActualStocks = x.ActualStocks!,
                     Ratings = x.Ratings!,
-                    AuthorIds = x.BookAuthors!.Select(ba => ba.AuthorId).ToList()
+                    AuthorIds = x.BookAuthors!.Select(ba => ba.AuthorId).ToList(),
+                    Images = x.BookImages.Select(b => b.ImageUrl).ToList(),
                 })
                 .ToListAsync();
 
@@ -59,32 +66,35 @@ namespace OnlineLibraryManagementSystem.Controllers
         }
 
 
-        //AddBook
         [HttpPost]
         [Route("addbook")]
-        public async Task<IActionResult> AddBook([FromBody] BookVM addBookVM)
+        public async Task<IActionResult> AddBook([FromForm] BookVM addBookVM, List<IFormFile> images)
         {
             try
             {
-                if (addBookVM == null) { return BadRequest(); }
+                if (addBookVM == null)
+                {
+                    return BadRequest();
+                }
+
                 if (!ModelState.IsValid)
                 {
                     return BadRequest();
                 }
 
-                // Check if book name already exists in database
+                // Check if book name already exists in the database
                 var existingBook = await _context.Books.FirstOrDefaultAsync(b => b.BookName == addBookVM.BookName);
-                var PublisherExists = await _context.Publisher.FirstOrDefaultAsync(p => p.Id == addBookVM.PublisherId);
+                // Check if publisher id is present or not in the database
+                var publisherExists = await _context.Publisher.FirstOrDefaultAsync(p => p.Id == addBookVM.PublisherId);
 
                 if (existingBook != null)
                 {
                     return Conflict($"Book with name '{addBookVM.BookName}' already exists.");
                 }
-                else if(PublisherExists == null)
+                else if (publisherExists == null)
                 {
-                    return NotFound($"Publisher with Id = '{addBookVM.PublisherId}' doesnot exists.");
+                    return NotFound($"Publisher with Id = '{addBookVM.PublisherId}' does not exist.");
                 }
-                
 
                 var book = new Book()
                 {
@@ -99,11 +109,14 @@ namespace OnlineLibraryManagementSystem.Controllers
                     Description = addBookVM.Description,
                     ActualStocks = addBookVM.ActualStocks,
                     Ratings = addBookVM.Ratings,
-                    BookAuthors = new List<BookAuthor>()
+                    BookAuthors = new List<BookAuthor>(),
+                    BookImages = new List<BookImage>()
                 };
+
                 foreach (var authorId in addBookVM.AuthorIds!)
                 {
                     var author = await _context.Authors.FindAsync(authorId);
+
                     if (author != null)
                     {
                         book.BookAuthors.Add(new BookAuthor()
@@ -113,11 +126,23 @@ namespace OnlineLibraryManagementSystem.Controllers
                     }
                     else
                     {
-                        return NotFound($"Author Id '{authorId}' doesnot exists.");
+                        return NotFound($"Author Id '{authorId}' does not exist.");
                     }
                 }
+
+                foreach (var image in images)
+                {
+                    var imageUrl = await UploadImageAsync(image, addBookVM.BookName);
+
+                    book.BookImages.Add(new BookImage()
+                    {
+                        ImageUrl = imageUrl!
+                    });
+                }
+
                 _context.Books.Add(book);
                 await _context.SaveChangesAsync();
+
                 return Ok(addBookVM);
             }
             catch (Exception ex)
@@ -126,95 +151,161 @@ namespace OnlineLibraryManagementSystem.Controllers
             }
         }
 
-
-        //UpdateBooksWithAuthorIdByBookId
-        [HttpPut]
-        [Route("updateBook")]
-        public async Task<IActionResult> UpdateBooksWithAuthorIdByBookId([FromQuery] int id, [FromBody] BookVM vm)
+        [NonAction]
+        private async Task<string?> UploadImageAsync(IFormFile image, string bname)
         {
-            var bookid = await _context.Books.FirstOrDefaultAsync(b => b.Id == id);
-
-            if (bookid == null)
+            if (image != null || image!.Length > 0)
             {
-                return NotFound("Book not found.");
-            }
+                string path = _webHostEnvironment.WebRootPath + "\\bookImages\\" + bname + "\\";
+                string fileName = image.FileName;
 
-            var PublisherExists = await _context.Publisher.FirstOrDefaultAsync(p => p.Id == vm.PublisherId);
-
-            if (PublisherExists == null)
-            {
-                return NotFound($"Publisher with Id = '{vm.PublisherId}' doesnot exists.");
-            }
-
-            int bookId = bookid!.Id;
-
-            var book = await _context.Books
-                .Include(b => b.BookAuthors)
-                .FirstOrDefaultAsync(b => b.Id == bookId);
-
-            book!.BookName = vm.BookName;
-            book.Genre = vm.Genre;
-            book.PublisherId= vm.PublisherId;
-            book.PublishDate = vm.PublishDate;
-            book.Language = vm.Language;
-            book.Edition = vm.Edition;
-            book.BookCost = vm.BookCost;
-            book.NumberOfPages = vm.NumberOfPages;
-            book.Description = vm.Description;
-            book.ActualStocks = vm.ActualStocks;
-            
-            book.Ratings = vm.Ratings;
-
-            var authorIds = vm.AuthorIds!.ToList();
-
-            var bookAuthorsToDelete = book.BookAuthors!
-                .Where(ba => !authorIds.Contains(ba.AuthorId))
-                .ToList();
-
-            foreach (var bookAuthorToDelete in bookAuthorsToDelete)
-            {
-                book.BookAuthors!.Remove(bookAuthorToDelete);
-            }
-
-            var bookAuthorsToAdd = authorIds
-                .Where(aid => !book.BookAuthors!.Any(ba => ba.AuthorId == aid))
-                .Select(aid => new BookAuthor { BookId = book.Id, AuthorId = aid })
-                .ToList();
-
-            foreach (var bookAuthorToAdd in bookAuthorsToAdd)
-            {
-                var authorExists = await _context.Authors.FindAsync(bookAuthorToAdd.AuthorId);
-                if(authorExists != null)
+                if (!Directory.Exists(path))
                 {
-                    book.BookAuthors!.Add(bookAuthorToAdd);
+                    Directory.CreateDirectory(path);
                 }
-                else
-                {
-                    return NotFound($"Author Id = '{bookAuthorToAdd.AuthorId}' doesnot exits");
-                }    
-            }
 
+                var filePath = Path.Combine(path, fileName);
+
+                using (FileStream fileStream = System.IO.File.Create(filePath))
+                {
+                    await image.CopyToAsync(fileStream);
+                    fileStream.Flush();
+                    return filePath;
+                }
+            }
+            return null;
+        }
+
+        [HttpPut]
+        [Route("updatebook/{id}")]
+        public async Task<IActionResult> UpdateBook(int id, [FromForm] BookVM bookVM, List<IFormFile> images)
+        {
             try
             {
+                var bookExists = BookExists(id);
+
+                if (!bookExists)
+                {
+                    return NotFound($"Book with Id = '{id}' not found.");
+                }
+
+                var book = await _context.Books.Include(b => b.BookAuthors).Include(b => b.BookImages).FirstOrDefaultAsync(b => b.Id == id);
+
+                if (book == null)
+                {
+                    return NotFound($"Book with Id = '{id}' not found.");
+                }
+
+                var publisherExists = await _context.Publisher.FirstOrDefaultAsync(p => p.Id == bookVM.PublisherId);
+                if (!ModelState.IsValid)
+                {
+                    return BadRequest();
+                }
+
+                // Check if publisher id is present or not in the database
+                else if (publisherExists == null)
+                {
+                    return NotFound($"Publisher with Id = '{bookVM.PublisherId}' does not exist.");
+                }
+
+                book.BookName = bookVM.BookName;
+                book.Genre = bookVM.Genre;
+                book.PublisherId = bookVM.PublisherId;
+                book.PublishDate = bookVM.PublishDate;
+                book.Language = bookVM.Language;
+                book.Edition = bookVM.Edition;
+                book.BookCost = bookVM.BookCost;
+                book.NumberOfPages = bookVM.NumberOfPages;
+                book.Description = bookVM.Description;
+                book.ActualStocks = bookVM.ActualStocks;
+                book.Ratings = bookVM.Ratings;
+
+                // Update book authors
+                var existingAuthorIds = book.BookAuthors!.Select(ba => ba.AuthorId);
+                var newAuthorIds = bookVM.AuthorIds!.Except(existingAuthorIds);
+
+                foreach (var authorId in newAuthorIds)
+                {
+                    var author = await _context.Authors.FindAsync(authorId);
+
+                    if (author != null)
+                    {
+                        book.BookAuthors!.Add(new BookAuthor()
+                        {
+                            AuthorId = authorId,
+                        });
+                    }
+                    else
+                    {
+                        return NotFound($"Author Id '{authorId}' does not exist.");
+                    }
+                }
+
+                foreach (var bookAuthor in book.BookAuthors!.ToList())
+                {
+                    if (!bookVM.AuthorIds!.Contains(bookAuthor.AuthorId))
+                    {
+                        book.BookAuthors!.Remove(bookAuthor);
+                    }
+                }
+
+                // Update book images
+                foreach (var bookImage in book.BookImages)
+                {
+                    _context.BookImages.Remove(bookImage);
+                }
+
+                foreach (var image in images)
+                {
+                    var imageUrl = await UploadUpdatedImageAsync(image, bookVM.BookName);
+
+                    book.BookImages.Add(new BookImage()
+                    {
+                        ImageUrl = imageUrl!
+                    });
+                }
+
                 await _context.SaveChangesAsync();
-            }
-            catch (DbUpdateConcurrencyException)
-            {
-                if (!BookExists(bookId))
-                {
-                    return NotFound("Book not found.");
-                }
-                else
-                {
-                    throw;
-                }
+
+                return Ok(bookVM);
             }
             catch (Exception ex)
             {
-                return StatusCode(StatusCodes.Status500InternalServerError, $"Error updating book: {ex.Message}");
+                return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
             }
+        }
 
-            return Ok("Book updated successfully.");
+        int counter = 0;
+        [NonAction]
+        private async Task<string?> UploadUpdatedImageAsync(IFormFile image, string bname)
+        {
+            if (image != null || image!.Length > 0)
+            {
+                string path = _webHostEnvironment.WebRootPath + "\\bookImages\\" + bname + "\\";
+                string fileName = image.FileName;
+
+                if (!Directory.Exists(path))
+                {
+                    Directory.CreateDirectory(path);
+                }
+                else if (counter == 0)
+                {
+                    Directory.Delete(path, true);
+                    Directory.CreateDirectory(path);
+                    counter++;
+                }
+
+
+                var filePath = Path.Combine(path, fileName);
+
+                using (FileStream fileStream = System.IO.File.Create(filePath))
+                {
+                    await image.CopyToAsync(fileStream);
+                    fileStream.Flush();
+                    return filePath;
+                }
+            }
+            return null;
         }
 
         private bool BookExists(int id)
@@ -236,9 +327,14 @@ namespace OnlineLibraryManagementSystem.Controllers
                     return NotFound("Book not found.");
                 }
 
-                _context.Books.Remove(book);
-                await _context.SaveChangesAsync();
-                return Ok("Book deleted successfully.");
+                var deleteImages = DeleteUploadedImageAsync(book.BookName);
+                if (deleteImages != null)
+                {
+                    _context.Books.Remove(book);
+                    await _context.SaveChangesAsync();
+                    return Ok("Book deleted successfully.");
+                }
+                return BadRequest("Book cannot be deleted.");
             }
             catch (Exception ex)
             {
@@ -247,8 +343,26 @@ namespace OnlineLibraryManagementSystem.Controllers
         }
 
 
+        [NonAction]
+        private string? DeleteUploadedImageAsync(string bname)
+        {
+            if (bname != null || bname!.Length > 0)
+            {
+                string path = _webHostEnvironment.WebRootPath + "\\bookImages\\" + bname + "\\";
+
+                if (Directory.Exists(path))
+                {
+                    Directory.Delete(path, true);
+
+                    return "book deleted succesfully";
+                }
+            }
+            return null;
+        }
+
 
         //-------------------------------AUTHOR-------------------------------
+
 
         //GetAllAuthor
         [HttpGet]
@@ -470,7 +584,5 @@ namespace OnlineLibraryManagementSystem.Controllers
                 return StatusCode(500, $"An error occurred while deleting the publisher: {ex.Message}");
             }
         }
-
-
     }
 }
